Moved create payloads to utilities and strong typed some fields
