+++ .env.sample
@@ -14,6 +14,7 @@ REACT_APP_ENABLE_ABOUT=true
 REACT_APP_ENABLE_TEAMS=true
 REACT_APP_ENABLE_MDA_POINT=true
 REACT_APP_ENABLE_PRACTITIONERS=true
+REACT_APP_ENABLE_JURISDICTION_METADATA_UPLOAD=true
 REACT_APP_DISABLE_LOGIN_PROTECTION=false
 REACT_APP_SUPERSET_API_BASE=https://superset.reveal-stage.smartregister.org/ # notice the ending /
 REACT_APP_SUPERSET_API_ENDPOINT=slice
