+++ src/constants.tsx
@@ -153,3 +153,6 @@ export const SORT_BY_EFFECTIVE_PERIOD_START_FIELD = 'plan_effective_period_start
 /** Query Params */
 export const QUERY_PARAM_TITLE = 'title';
 export const QUERY_PARAM_USER = 'user';
+
+/** Settings Configuration */
+export const SETTINGS_CONFIGURATION = 'SettingConfiguration';
